custom-vision-import-export-material/requirements.txt
azure-cognitiveservices-vision-customvision
pillow